src/main/java/training/todo/shared/Utils.java
package training.todo.shared;

import org.springframework.stereotype.Component;

import java.security.SecureRandom;
import java.util.Random;

@Component
public class Utils {
    private final Random RANDOM = new SecureRandom();
    private final String ALPHABET = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";

    public String generateUserId(int length) {
        return generateRandomString(length);
    }

    // this function will generate a random string containing digits and alphabetical
    // characters (lowercase and uppercase)...
    private String generateRandomString(int length) {
        StringBuilder randomString = new StringBuilder(length);
        for(int i = 0; i < length; i++) {
            randomString.append(ALPHABET.charAt(RANDOM.nextInt(ALPHABET.length())));
        }

        return new String(randomString);
    }
}
